+++ src/model_signing/signing.py
@@ -121,6 +121,7 @@ class Config:
         oidc_issuer: Optional[str] = None,
         use_ambient_credentials: bool = False,
         use_staging: bool = False,
+        force_oob: bool = False,
         identity_token: Optional[str] = None,
     ) -> Self:
         """Configures the signing to be performed with Sigstore.
@@ -138,6 +139,11 @@ class Config:
               signer identity via OIDC will start.
             use_staging: Use staging configurations, instead of production. This
               is supposed to be set to True only when testing. Default is False.
+            force_oob: If True, forces an out-of-band (OOB) OAuth flow. If set,
+              the OAuth authentication will not attempt to open the default web
+              browser. Instead, it will display a URL and code for manual
+              authentication. Default is False, which means the browser will be
+              opened automatically if possible.
             identity_token: An explicit identity token to use when signing,
               taking precedence over any ambient credential or OAuth workflow.
 
@@ -149,6 +155,7 @@ class Config:
             use_ambient_credentials=use_ambient_credentials,
             use_staging=use_staging,
             identity_token=identity_token,
+            force_oob=force_oob,
         )
         return self